facebook user info

+++ app/src/main/java/com/joe/chill/structs/MatchCard.java
@@ -13,9 +13,9 @@ public class MatchCard implements Parcelable {
   private String mName;
   private String mBio;
   private List<String> mImageUrls;
-  private int mAge;
+  private long mAge;
 
-  public MatchCard(String userId, String name, String bio, int age, List<String> imageUrls) {
+  public MatchCard(String userId, String name, String bio, long age, List<String> imageUrls) {
     setUserId(userId);
     setName(name);
     setImageUrls(imageUrls);
@@ -26,7 +26,7 @@ public class MatchCard implements Parcelable {
   protected MatchCard(Parcel in) {
     mUserId = in.readString();
     mName = in.readString();
-    mAge = in.readInt();
+    mAge = in.readLong();
     mImageUrls = in.createStringArrayList();
     mBio = in.readString();
   }
@@ -75,11 +75,11 @@ public class MatchCard implements Parcelable {
     return mImageUrls;
   }
 
-  public void setAge(int age) {
+  public void setAge(long age) {
     mAge = age;
   }
 
-  public int getAge() {
+  public long getAge() {
     return mAge;
   }
 
@@ -92,7 +92,7 @@ public class MatchCard implements Parcelable {
   public void writeToParcel(Parcel dest, int flags) {
     dest.writeString(getUserId());
     dest.writeString(getName());
-    dest.writeInt(getAge());
+    dest.writeLong(getAge());
     dest.writeStringList(getImageUrls());
     dest.writeString(getUserBio());
   }
